Feat: 회원가입 인터페이스 완성 # #로 시작하는 줄은 주석입니다. 주석은 지우지 마세요!! # <Type>(<scope>): <제목> (가능하면 50자 이내, 마침표 없이) # -----type 목록----- # Feat: 새 기능 추가 # Fix: 버그 수정 # Docs: 문서 추가, 수정 등 # Refactor: 리팩터링(버그는 아니지만 코드가 너무 별로라서 수정하는 작업) # Test: 테스트 코드 추가 등 # Chore: 위 항목 중 해당되는 것이 없을 때
# 본문 (필요 시 작성, 가능하면 한 줄에 72자 이내)
# - 어떤 변경을 했는지
# - 왜 이 변경이 필요한지
# - 추가적으로 알릴 사항이 있다면 여기에

# Footer (이슈 번호 언급 등 등)
# 예: 이슈 #123 해결
