+++ .gitmessage
@@ -1,4 +1,4 @@
-Feat: 로그인 과정에서 사용자 경험 향상
+Feat: 회원가입 인터페이스 완성
 # #로 시작하는 줄은 주석입니다. 주석은 지우지 마세요!!
 # <Type>(<scope>): <제목> (가능하면 50자 이내, 마침표 없이)
 # -----type 목록-----
